+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## HEAD (Unreleased)
 
-**(none)**
+- feat: Add `run-program` flag support for refresh, destroy, and preview
+  commands ([#1376](https://github.com/pulumi/actions/pull/1376))
 
 ---